Rename GiveArgument to StackAmountArgument
Renamed and moved GiveArgument -> StackArgument in litecommand/argument/StackAmountArgument

Took 4 minutes
