+++ eternalcore-core/src/main/java/com/eternalcode/core/feature/give/GiveCommand.java
@@ -3,6 +3,7 @@ package com.eternalcode.core.feature.give;
 import com.eternalcode.annotations.scan.command.DescriptionDocs;
 import com.eternalcode.core.configuration.implementation.PluginConfiguration;
 import com.eternalcode.core.injector.annotations.Inject;
+import com.eternalcode.core.litecommand.argument.StackAmountArgument;
 import com.eternalcode.core.notice.NoticeService;
 import com.eternalcode.core.util.MaterialUtil;
 import dev.rollczi.litecommands.annotations.argument.Arg;
@@ -44,7 +45,7 @@ class GiveCommand {
 
     @Execute
     @DescriptionDocs(description = "Gives you an item with a custom amount", arguments = "<item> [amount]")
-    void execute(@Context Player player, @Arg Material material, @OptionalArg(GiveArgument.KEY) int amount) {
+    void execute(@Context Player player, @Arg Material material, @OptionalArg(StackAmountArgument.KEY) int amount) {
         boolean isSuccess = this.giveService.giveItem(player, player, material, amount);
 
         if (isSuccess) {
@@ -58,7 +59,7 @@ class GiveCommand {
 
     @Execute
     @DescriptionDocs(description = "Gives an item with a custom amount to another player", arguments = "<player> <item> <amount>")
-    void execute(@Context CommandSender sender, @OptionalArg Player target, @Arg Material material, @OptionalArg(GiveArgument.KEY) int amount) {
+    void execute(@Context CommandSender sender, @OptionalArg Player target, @Arg Material material, @OptionalArg(StackAmountArgument.KEY) int amount) {
         boolean isSuccess = this.giveService.giveItem(sender, target, material, amount);
 
         if (!isSuccess) {
