+++ eternalcore-core/src/main/java/com/eternalcode/core/litecommand/LiteCommandsSetup.java
@@ -14,7 +14,6 @@ import dev.rollczi.litecommands.annotations.LiteCommandsAnnotations;
 import dev.rollczi.litecommands.bukkit.LiteBukkitFactory;
 import dev.rollczi.litecommands.bukkit.LiteBukkitMessages;
 import dev.rollczi.litecommands.folia.FoliaExtension;
-import dev.rollczi.litecommands.schematic.SchematicFormat;
 import net.kyori.adventure.platform.AudienceProvider;
 import net.kyori.adventure.text.minimessage.MiniMessage;
 import org.bukkit.Server;
