+++ eternalcore-core/src/main/java/com/eternalcode/core/litecommand/argument/StackAmountArgument.java
@@ -1,6 +1,5 @@
-package com.eternalcode.core.feature.give;
+package com.eternalcode.core.litecommand.argument;
 
-import com.eternalcode.core.litecommand.argument.AbstractViewerArgument;
 import com.eternalcode.core.injector.annotations.Inject;
 import com.eternalcode.core.injector.annotations.lite.LiteArgument;
 import com.eternalcode.core.translation.Translation;
@@ -14,14 +13,14 @@ import org.bukkit.command.CommandSender;
 
 import java.util.List;
 
-@LiteArgument(type = int.class, name = GiveArgument.KEY)
-public class GiveArgument extends AbstractViewerArgument<Integer> {
+@LiteArgument(type = int.class, name = StackAmountArgument.KEY)
+public class StackAmountArgument extends AbstractViewerArgument<Integer> {
 
     private static final List<Integer> suggestions = List.of(1, 8, 16, 32, 64);
     public static final String KEY = "amount";
 
     @Inject
-    public GiveArgument(TranslationManager translationManager) {
+    public StackAmountArgument(TranslationManager translationManager) {
         super(translationManager);
     }
